try plotting
Former-commit-id: f343bb1b7f57c49cd87e6cc2fbaacd078d11b87c
Former-commit-id: 08e4e98ea3b8197ad503d5a751a2f454a54d148f
Former-commit-id: e2d121839ad9aefb2828ba313d4c70cf4c6487a1
Former-commit-id: 433545bc30ffdf942821203c75256840e74861eb
Former-commit-id: b840c4bff38a1d29cd89d750a9bf8b3a346b0ac2 [formerly 7483d3e823be568ac7b8a2aa960afb7d84d308b2]
Former-commit-id: cd9d49da2d2327530db13bef92ff110d9d09e37b

+++ SimulateNetwork_extra_funcs.py
@@ -417,7 +417,7 @@ def single_run_numba(N0, mu, alpha, psi, beta, sigma, Ninit, Mrate1, Mrate2, gam
                     inner = []
                     for jx in range(mm):
                         if AK[ix, jx] > -1:
-                            inner.append(AK[ix, jx])
+                            inner.append(int(AK[ix, jx]))
                     outer.append(inner)
                 SIRfile_AK.append(outer)
 
@@ -539,7 +539,7 @@ def single_run_and_save(filename):
         joblib.dump([SIRfile_SK, SIRfile_P1, SIRfile_UK], filename_SK_P1_UK)
         # pickle.dump([SIRfile_SK, SIRfile_P1, SIRfile_UK], open(filename_SK_P1_UK.replace('joblib', 'pickle'), "wb"))
 
-        SIRfile_AK = awkward.fromiter(SIRfile_AK)
+        SIRfile_AK = awkward.fromiter(SIRfile_AK).astype(np.int32)
         filename_AK = filename_SK_P1_UK.replace('SK_P1_UK.joblib', 'AK.parquet')
         awkward.toparquet(filename_AK, SIRfile_AK)
